server/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "store-system-server"
version = "1.0.0"
description = "Store System Server - A comprehensive retail management system"
readme = "README.md"
authors = [
    {name = "Store System Team"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.110.1",
    "uvicorn==0.29.0",
    "psycopg2==2.9.9",
    "requests==2.31.0",
    "httpx==0.27.0",
    "python-dotenv==1.0.1",
    "xlrd==2.0.1",
    "openpyxl==3.1.2",
    "python_multipart==0.0.9",
    "PyJWT==2.8.0",
    "bcrypt==4.1.3",
    "pandas==2.2.2",
    "numpy==2.0.1",
    "types_psycopg2==2.9.21.20250121",
    "pandas-stubs>=2.2.0",
    "xgboost==2.0.3",
    "scikit-learn==1.4.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "black",
    "flake8",
    "mypy",
]

[tool.setuptools]
packages = ["store_system_server"]

[tool.setuptools.package-dir]
store_system_server = "."
